refactor: change count type from float to long in stats analytics

+++ gravitee-apim-repository/gravitee-apim-repository-elasticsearch/src/main/java/io/gravitee/repository/elasticsearch/v4/analytics/adapter/StatsQueryAdapter.java
@@ -99,7 +99,7 @@ public class StatsQueryAdapter {
             return Optional.empty();
         }
 
-        float count = agg.getCount();
+        long count = agg.getCount().longValue();
         float sum = agg.getSum();
         float avg = agg.getAvg();
         float min = agg.getMin();
